requirements.txt
fastapi 
uvicorn 
jinja2 
qrcode 
pillow 
python-barcode
uvicorn